requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
python-dotenv==1.0.1
psycopg2-binary==2.9.9
tweepy==4.14.0
praw==7.7.1
APScheduler==3.10.4
openai==1.46.0
python-telegram-bot==21.6
httpx==0.27.2
tenacity==9.0.0
orjson==3.10.7
pydantic-settings==2.11.0
slowapi==0.1.9
prometheus-client==0.20.0
prometheus-fastapi-instrumentator==7.0.0
pytest==8.3.3
